requirements.txt
# Core Deep Learning Framework
tensorflow>=2.13.0
tf-keras>=2.13.0

# Transformers for BERT
transformers>=4.30.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
openai-whisper>=20231117

# Video Processing
moviepy>=1.0.3

# 3D Point Cloud Processing
#open3d-python>=0.19
trimesh>=3.22.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Utilities
pillow>=10.0.0
